fix: Tests after module rename
